requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.4
djangorestframework==3.15.2
docopt==0.6.2
et_xmlfile==2.0.0
fastapi==0.115.6
fonttools==4.55.3
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
num2words==0.5.14
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.34.0
weasyprint==63.1
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.18.3
zopfli==0.2.3.post1
